+++ src/main/java/subway/presentation/request/StationCreateRequest.java
@@ -1,7 +1,10 @@
 package subway.presentation.request;
 
+import javax.validation.constraints.NotBlank;
+
 public class StationCreateRequest {
 
+    @NotBlank(message = "역을 입력해주세요.")
     private String name;
 
     private StationCreateRequest() {
